Project 0 Cs50x
This Scratch project features a dog and a star sprite. When the dog sprite touches the star sprite, a random click sound and a random dog sound are played. The sounds stop when the dog sprite leaves the star sprite. To use the project, simply click the green flag to start it, and use the arrow keys to move the dog sprite around the screen. When the dog sprite touches the star sprite, you will hear a random click sound and a random dog sound. When the dog sprite moves away from the star sprite, the sounds will stop. This project demonstrates the use of conditional statements and random number generators in Scratch.